Refactor polaris mutators
