+++ pkg/polaris/creater.go
@@ -22,11 +22,14 @@
 package polaris
 
 import (
+	"encoding/json"
 	"fmt"
 
 	routev1 "github.com/openshift/api/route/v1"
 	securityv1 "github.com/openshift/api/security/v1"
+	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/meta"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/client-go/kubernetes/scheme"
@@ -38,6 +41,10 @@ import (
 func GetComponents(baseURL string, polaris Polaris) (map[string]runtime.Object, error) {
 	components := make(map[string]runtime.Object)
 
+	baseSecrets, err := GetPolarisBaseSecrets(polaris)
+	if err != nil {
+		return nil, err
+	}
 	dbComponents, err := GetPolarisDBComponents(baseURL, polaris)
 	if err != nil {
 		return nil, err
@@ -47,6 +54,9 @@ func GetComponents(baseURL string, polaris Polaris) (map[string]runtime.Object, 
 		return nil, err
 	}
 
+	for k, v := range baseSecrets {
+		components[k] = v
+	}
 	for k, v := range dbComponents {
 		components[k] = v
 	}
@@ -109,4 +119,142 @@ func ConvertYamlFileToRuntimeObjects(stringContent string) map[string]runtime.Ob
 		mapOfUniqueIDToDesiredRuntimeObject[uniqueID] = runtimeObject
 	}
 	return mapOfUniqueIDToDesiredRuntimeObject
+}
+
+func GetPolarisBaseSecrets(polaris Polaris) (map[string]runtime.Object, error) {
+	mapOfUniqueIDToBaseRuntimeObject := make(map[string]runtime.Object, 0)
+
+	// Coverity license
+	mapOfUniqueIDToBaseRuntimeObject["Secret.coverity-license"] = &corev1.Secret{
+		TypeMeta: metav1.TypeMeta{
+			Kind:       "Secret",
+			APIVersion: "v1",
+		},
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      "coverity-license",
+			Namespace: polaris.Namespace,
+			Labels: map[string]string{
+				"environment": polaris.Namespace,
+			},
+		},
+		Data: map[string][]byte{
+			"license": []byte(polaris.Licenses.Coverity),
+		},
+		Type: corev1.SecretTypeOpaque,
+	}
+
+	// Tool store sync service account
+	mapOfUniqueIDToBaseRuntimeObject["Secret.tools-store-sync"] = &corev1.Secret{
+		TypeMeta: metav1.TypeMeta{
+			Kind:       "Secret",
+			APIVersion: "v1",
+		},
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      "tools-store-sync",
+			Namespace: polaris.Namespace,
+			Labels: map[string]string{
+				"environment": polaris.Namespace,
+			},
+		},
+		Data: map[string][]byte{
+			"credentials.json": []byte(polaris.GCPServiceAccount),
+			"instance-name":    []byte("tools-store-sync"),
+		},
+		Type: corev1.SecretTypeOpaque,
+	}
+
+	// Pull secret gcr-json-key
+	type dockerAuthConfig struct {
+		Username string
+		Password string
+		Email    string
+	}
+
+	dockerCfg := struct {
+		Auths map[string]dockerAuthConfig `json:"auths"`
+	}{
+		Auths: map[string]dockerAuthConfig{
+			"https://gcr.io": {
+				Username: "_json_key",
+				Password: polaris.GCPServiceAccount,
+			},
+		},
+	}
+
+	dockerCfgByte, err := json.Marshal(dockerCfg)
+	if err != nil {
+		return nil, err
+	}
+
+	mapOfUniqueIDToBaseRuntimeObject["Secret.gcr-json-key"] = &corev1.Secret{
+		TypeMeta: metav1.TypeMeta{
+			Kind:       "Secret",
+			APIVersion: "v1",
+		},
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      "gcr-json-key",
+			Namespace: polaris.Namespace,
+			Labels: map[string]string{
+				"environment": polaris.Namespace,
+			},
+		},
+		Data: map[string][]byte{
+			".dockerconfigjson": dockerCfgByte,
+		},
+		Type: corev1.SecretTypeDockerConfigJson,
+	}
+
+	return mapOfUniqueIDToBaseRuntimeObject, nil
+}
+
+// GetPolarisReportingComponents get Polaris reporting components
+func GetPolarisReportingComponents(baseURL string, polaris Polaris) (map[string]runtime.Object, error) {
+	content, err := GetBaseYaml(baseURL, "polaris", polaris.Version, "reporting_base.yaml")
+	if err != nil {
+		return nil, err
+	}
+
+	return fromYaml(content, polaris)
+}
+
+// GetPolarisDBComponents get Polaris DB components
+func GetPolarisDBComponents(baseURL string, polaris Polaris) (map[string]runtime.Object, error) {
+	content, err := GetBaseYaml(baseURL, "polaris", polaris.Version, "polarisdb_base.yaml")
+	if err != nil {
+		return nil, err
+	}
+
+	return fromYaml(content, polaris)
+}
+
+// GetPolarisComponents get Polaris components
+func GetPolarisComponents(baseURL string, polaris Polaris) (map[string]runtime.Object, error) {
+	content, err := GetBaseYaml(baseURL, "polaris", polaris.Version, "polaris_base.yaml")
+	if err != nil {
+		return nil, err
+	}
+
+	return fromYaml(content, polaris)
+}
+
+// GetPolarisProvisionComponents get Polaris provision components
+func GetPolarisProvisionComponents(baseURL string, polaris Polaris) (map[string]runtime.Object, error) {
+	content, err := GetBaseYaml(baseURL, "polaris", polaris.Version, "organization-provision-job.yaml")
+	if err != nil {
+		return nil, err
+	}
+
+	return fromYaml(content, polaris)
+}
+
+func removeTestManifests(objects map[string]runtime.Object) map[string]runtime.Object {
+	objectsToBeRemoved := []string{
+		"Pod.swip-db-ui-test-off9s",
+		"Pod.swip-db-vault-status-test",
+		"ConfigMap.polaris-db-consul-tests",
+	}
+	for _, object := range objectsToBeRemoved {
+		delete(objects, object)
+	}
+	return objects
 }
